requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.7.2
python-chess==1.999
python-dotenv==1.0.0
gunicorn==21.2.0